// src/@swagger/model/filter.ts
/**
 * Simplified Accounting System (SAS)
 * <h3>How to generate API clients and server stubs:</h3>java -jar <a href='http://119.15.88.122:8989/swagger-codegen-cli-2.3.1.jar' target='_blank'>swagger-codegen-cli-2.3.1.jar</a> generate -l typescript-angular -i http://119.15.88.122:8989/v2/api-docs -o /Users/sangkhim/Desktop/swagger<h3>LANGUAGES:</h3>-l typescript-angular (Angular)<br />-l javascript (React)<br />-l java (Android)<br />-l kotlin (Android)<br />-l objc (Objective C)<br />-l swift4 (Swift 4)
 *
 * OpenAPI spec version: API
 * Contact: info@udaya-tech.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Filter {
    page?: number;
    rowsPerPage?: number;
    searchText?: string;
}